+++ include/swift/Basic/ColorUtils.h
@@ -10,7 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 //
-// This file defines an 'OSColor' class for helping printing colorful outputs
+// This file defines several utilities for helping print colorful outputs
 // to the terminal.
 //
 //===----------------------------------------------------------------------===//
@@ -40,6 +40,59 @@ public:
 
   OSColor &operator<<(char C) { OS << C; return *this; }
   OSColor &operator<<(llvm::StringRef Str) { OS << Str; return *this; }
+};
+
+/// A stream which forces color output.
+class ColoredStream : public raw_ostream {
+  raw_ostream &Underlying;
+
+public:
+  explicit ColoredStream(raw_ostream &underlying) : Underlying(underlying) {}
+  ~ColoredStream() override { flush(); }
+
+  raw_ostream &changeColor(Colors color, bool bold = false,
+                           bool bg = false) override {
+    Underlying.changeColor(color, bold, bg);
+    return *this;
+  }
+  raw_ostream &resetColor() override {
+    Underlying.resetColor();
+    return *this;
+  }
+  raw_ostream &reverseColor() override {
+    Underlying.reverseColor();
+    return *this;
+  }
+  bool has_colors() const override { return true; }
+
+  void write_impl(const char *ptr, size_t size) override {
+    Underlying.write(ptr, size);
+  }
+  uint64_t current_pos() const override {
+    return Underlying.tell() - GetNumBytesInBuffer();
+  }
+
+  size_t preferred_buffer_size() const override { return 0; }
+};
+
+/// A stream which drops all color settings.
+class NoColorStream : public raw_ostream {
+  raw_ostream &Underlying;
+
+public:
+  explicit NoColorStream(raw_ostream &underlying) : Underlying(underlying) {}
+  ~NoColorStream() override { flush(); }
+
+  bool has_colors() const override { return false; }
+
+  void write_impl(const char *ptr, size_t size) override {
+    Underlying.write(ptr, size);
+  }
+  uint64_t current_pos() const override {
+    return Underlying.tell() - GetNumBytesInBuffer();
+  }
+
+  size_t preferred_buffer_size() const override { return 0; }
 };
 
 } // namespace swift
